Total usage calculation
ATS in info.plist

+++ SG Mobile Usage/Models/QuarterUsage.swift
@@ -10,9 +10,9 @@ import Foundation
 
 struct QuarterUsage {
     
-    private(set) var mobileDataVolume: Double, quarter: String, year: Int8, id: Int
+    private(set) var mobileDataVolume: Double, quarter: String, year: Int, id: Int
     
-    init(id: Int, year: Int8, quarter: String, volume: Double) {
+    init(id: Int, year: Int, quarter: String, volume: Double) {
         self.id = id
         self.mobileDataVolume = volume
         self.year = year
@@ -22,13 +22,13 @@ struct QuarterUsage {
     init(json: [String: Any]) {
         let id = json[Constants.JSONKey.id] as? Int ?? -1
         let dataVolume = Double(json[Constants.JSONKey.volume] as? String ?? "0.0")!
-        var year: Int8 = 0
+        var year: Int = 0
         var quarter = ""
         
         if let quarterStr = json[Constants.JSONKey.quarter] as? String {
             let components = quarterStr.split(separator: "-")
             if(components.count == 2) {
-                year = Int8(components.first ?? "0") ?? Int8(0)
+                year = Int(components.first ?? "0") ?? 0
                 quarter = String(components.last ?? "")
             }
         }
